backend/requirements.txt
streamlit==1.28.1
openai>=1.97.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
typing-extensions>=4.11
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.77 